Ensure serial output code match the intention of the developer

+++ Marlin/src/gcode/calibrate/M48.cpp
@@ -240,7 +240,7 @@ void GcodeSuite::M48() {
       sigma = SQRT(dev_sum / (n + 1));
 
       if (verbose_level > 1) {
-        SERIAL_ECHO((int)(n + 1));
+        SERIAL_ECHO(n + 1);
         SERIAL_ECHOPAIR(" of ", (int)n_samples);
         SERIAL_ECHOPAIR_F(": z: ", pz, 3);
         SERIAL_CHAR(' ');
